Address review feedback on the machines PR
- createMachine now sends virtualization=false explicitly instead of dropping the key and falling back to the system default
- extract MachineHomeMount to a shared top-level type rather than nesting it in CreateMachineView
- guard the stop-reason check and log-viewer fetch against stale async writes when the target changes mid-flight
- drop the outdated 'M3 absent' note on the MachineBackend protocol
